requirements.txt
Flask
tensorflow
transformers
Pillow
requests
pandas
